requirements.txt
alembic==1.13.2
Babel==2.13.0
blinker==1.8.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
<<<<<<< Updated upstream
=======
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
>>>>>>> Stashed changes
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
ghp-import==2.1.0
idna==2.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
Markdown==3.5
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.5.3
# Editable install with no version control (mkdocs-material==9.4.5)
-e /Users/valentincardozo/Projects/valencardozo.github.io/mkdocs-material
mkdocs-material-extensions==1.2
packaging==23.2
paginate==0.5.6
pathspec==0.11.2
platformdirs==3.11.0
Pygments==2.16.1
pymdown-extensions==10.3
PyMySQL==1.1.1
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
python-dateutil==2.8.2
PyYAML @ file:///private/tmp/pyyaml-20230717-6270-1iiwi1j/PyYAML-6.0.1
pyyaml_env_tag==0.1
regex==2023.10.3
requests==2.20.0
shiboken6==6.5.2
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==1.24.2
watchdog==3.0.0
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.19.2
